backend/requirements.txt
fastapi
uvicorn
python-multipart
motor
openai-whisper
ffmpeg-python
pydub
python-dotenv
requests
webrtcvad
numpy
scipy
torch
